% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{cyto_install_deps}
\alias{cyto_install_deps}
\title{Install package dependencies if needed.}
\usage{
cyto_install_deps(
  pkgdir = ".",
  dependencies = NA,
  repos = getOption("repos"),
  type = getOption("pkgType"),
  upgrade = c("default", "ask", "always", "never"),
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  build_vignettes = FALSE,
  ...
)
}
\arguments{
\item{build}{If `TRUE` build the package before installing.}

\item{build_opts}{Options to pass to `R CMD build`, only used when `build`}

\item{build_manual}{If `FALSE`, don't build PDF manual ('--no-manual').}

\item{build_vignettes}{If `FALSE`, don't build package vignettes ('--no-build-vignettes').
is `TRUE}

\item{...}{additional arguments passed to [utils::install.packages()].}
}
\description{
Install package dependencies if needed.
}
\examples{
\dontrun{
cyto_install_deps("../ggcyto")
}
}
