% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{cyto_pkg_deps}
\alias{cyto_pkg_deps}
\title{Find all dependencies of a cyto package.}
\usage{
cyto_pkg_deps(pkdir = ".", bioc_ver = bioc_version(), ...)
}
\arguments{
\item{pkdir}{path to a package directory , or to a package tarball}

\item{...}{arguments passed to [remotes::dev_package_deps()]}
}
\description{
Find all the dependencies of a package and determine whether they are ahead
or behind CRAN, bioc and CYTO repos.
}
\examples{
\dontrun{
cyto_pkg_deps("../ggcyto/")
}
}
