% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cyto.R
\name{cyto_pkg_github_url}
\alias{cyto_pkg_github_url}
\title{get the download url of the package hosted as github release assets}
\usage{
cyto_pkg_github_url(
  pkg,
  owner = getOption("cyto_repo_owner"),
  bioc_ver = bioc_version()
)
}
\arguments{
\item{pkg}{the package name}

\item{owner}{the github owner}

\item{bioc_ver}{the corresponding bioconductor version}
}
\description{
get the download url of the package hosted as github release assets
}
\examples{
\dontrun{
cyto_pkg_github_url("ggcyto")
}
}
