% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cyto.R
\name{cyto_remote}
\alias{cyto_remote}
\title{construct a 'cyto_remote' object}
\usage{
cyto_remote(
  pkg,
  owner = getOption("cyto_repo_owner"),
  type = getOption("pkgType"),
  ...
)
}
\arguments{
\item{pkg}{package name}

\item{owner}{github repo owner}
}
\description{
It queries the remote github repo and fetch the packager versions and download url
}
\examples{
remote <- cyto_remote("ggcyto")
remote_package_name(remote)
}
