% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{desc_to_local}
\alias{desc_to_local}
\title{Download DESCRIPTION file of the pkg from cyto repo}
\usage{
desc_to_local(pkg)
}
\arguments{
\item{pkg}{package name.}
}
\description{
Download DESCRIPTION file of the pkg from cyto repo
}
