% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cyto.R
\name{install_cyto}
\alias{install_cyto}
\title{Attempts to install a package from cyto repo}
\usage{
install_cyto(
  pkg,
  type = getOption("pkgType"),
  dependencies = NA,
  upgrade = c("default", "ask", "always", "never"),
  force = FALSE,
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  build_vignettes = FALSE,
  ...
)
}
\arguments{
\item{pkgs}{Character vector of packages to install.}
}
\description{
This function is vectorised on `pkgs` so you can install multiple
packages in a single command.
}
\examples{
\dontrun{
install_cyto("ggcyto")
}
}
\concept{package installation}
