% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cyto.R
\name{cyto_remote}
\alias{cyto_remote}
\title{construct a 'cyto_remote' object}
\usage{
cyto_remote(pkg, type = getOption("pkgType"), ...)
}
\arguments{
\item{...}{additional arguments passed to [utils::install.packages()].}
}
\description{
It queries the remote github repo and fetch the packager versions and download url
}
\examples{
\dontrun{
cyto_remote("ggcyto")
}
}
