% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_cyto.R
\name{install_cyto}
\alias{install_cyto}
\title{Attempts to install a package from cyto repo}
\usage{
install_cyto(
  pkg,
  type = getOption("pkgType"),
  dependencies = NA,
  upgrade = c("default", "ask", "always", "never"),
  force = FALSE,
  quiet = FALSE,
  build = TRUE,
  build_opts = c("--no-resave-data", "--no-manual", "--no-build-vignettes"),
  build_manual = FALSE,
  build_vignettes = FALSE,
  ...
)
}
\arguments{
\item{build}{If `TRUE` build the package before installing.}

\item{build_opts}{Options to pass to `R CMD build`, only used when `build`}

\item{build_manual}{If `FALSE`, don't build PDF manual ('--no-manual').}

\item{build_vignettes}{If `FALSE`, don't build package vignettes ('--no-build-vignettes').
is `TRUE}

\item{...}{additional arguments passed to [utils::install.packages()].}
}
\description{
This function is vectorised on `pkgs` so you can install multiple
packages in a single command.
}
\examples{
\dontrun{
install_cyto("ggcyto")
}
}
\concept{package installation}
